requirements.txt
azure-cosmos
azure-ai-ml
azure-ai-resources
azure-search-documents==11.4.0
openai
python-dotenv